refactor: 중복되는 코드를 `Piece`로 이동
